services/activity/pyproject.toml
[project]
name = "activity"
version = "0.1.0"
description = "Activity Service"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2",
    "django-ninja>=1.0",
    "psycopg[binary]",
    "dj-database-url",
    "gunicorn",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app", "src/activity", "src/core"]
